requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-generator==1.10
attrs==24.2.0
certifi==2024.8.30
certipy==0.1.3
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
distro==1.9.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
ipython-genutils==0.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpointer==2.0
jupyter_vscode_proxy==0.6
kiwisolver==1.4.7
lightning-utilities==0.11.8
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
openai==1.52.2
packaging==24.1
pamela==1.0.0
pandas==2.2.3
pdfminer.six==20240706
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
pycurl==7.45.1
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.12
pyOpenSSL==24.2.1
pyparsing==3.2.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
rdkit==2024.3.5
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.0
torchmetrics==1.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
yarl==1.16.0
zstandard==0.19.0
